+++ babel.config.js
@@ -1,9 +1,0 @@
-/* global process module */
-
-module.exports = (api) => {
-  api.cache(() => process.env.NODE_ENV);
-
-  return {
-    presets: ['@babel/env'],
-  };
-};