+++ jest.config.js
@@ -1,16 +1,8 @@
 /* global module */
 
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  globals: {
-    'ts-jest': {},
-  },
-  moduleNameMapper: {
-    '@kofein/(.+)$': '<rootDir>packages/$1/lib',
-  },
   transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
+    '^.+\\.(ts|tsx)$': '@swc/jest',
   },
   testPathIgnorePatterns: ['.*\\.js'],
 };