+++ lib/index.ts
@@ -1,6 +1,0 @@
-import Event, { Handler } from './Event';
-import ReadonlyEvent from './ReadonlyEvent';
-
-export default Event;
-export { ReadonlyEvent };
-export type { Handler };