+++ src/aggregateError.ts
@@ -1,0 +1,5 @@
+export class AggregateError extends Error {
+  constructor(readonly errors: unknown[]) {
+    super('Some handlers has errors');
+  }
+}