+++ src/index.ts
@@ -1,0 +1,4 @@
+export * from './baseEvent';
+export * from './event';
+export * from './asyncEvent';
+export * from './aggregateError';